work in progress on functions
