+++ ui/apps/dashboard/src/components/Functions/ActionMenu.tsx
@@ -3,11 +3,7 @@
 import Link from 'next/link';
 import { Select } from '@inngest/components/Select/Select';
 
-const StatusIcon = ({ className }: { className: string }) => (
-  <span className={`block h-2 w-2 shrink-0 rounded-full ${className}`} />
-);
-
-export const StatusMenu = ({ envSlug, archived }: { envSlug: string; archived: boolean }) => {
+export const ActionsMenu = ({ envSlug, archived }: { envSlug: string; archived: boolean }) => {
   const activeOption = { id: 'active', name: 'Active apps' };
   const archivedOption = { id: 'archived', name: 'Archived apps' };
   return (
@@ -21,15 +17,13 @@ export const StatusMenu = ({ envSlug, archived }: { envSlug: string; archived: b
     >
       <Select.Button className="w-[124px] px-4">
         <div className="text-basis mr-2 flex flex-row items-center text-xs font-medium leading-tight">
-          <StatusIcon className={`mr-2 ${archived ? 'bg-accent-subtle' : 'bg-primary-moderate'}`} />
-          {archived ? 'Archived' : 'Active'}
+          All actions
         </div>
       </Select.Button>
       <Select.Options>
         <Link href={`/env/${envSlug}/apps`} prefetch={true}>
           <Select.Option key={activeOption.id} option={activeOption}>
             <div className="text-basis flex flex-row items-center text-xs font-medium">
-              <StatusIcon className="bg-primary-moderate mr-2" />
               {activeOption.name}
             </div>
           </Select.Option>
@@ -37,7 +31,6 @@ export const StatusMenu = ({ envSlug, archived }: { envSlug: string; archived: b
         <Link href={`/env/${envSlug}/apps?archived=true`} prefetch={true}>
           <Select.Option key={archivedOption.id} option={archivedOption}>
             <div className="text-basis flex flex-row items-center text-xs font-medium">
-              <StatusIcon className="bg-accent-subtle mr-2" />
               {archivedOption.name}
             </div>
           </Select.Option>
