+++ ui/apps/dashboard/src/components/Functions/PauseFunction/PauseModal.tsx
@@ -1,15 +1,11 @@
 'use client';
 
 import { useState } from 'react';
-import { Button } from '@inngest/components/Button';
 import { AlertModal } from '@inngest/components/Modal';
 import { Select } from '@inngest/components/Select/Select';
-import { Tooltip, TooltipContent, TooltipTrigger } from '@inngest/components/Tooltip';
-import { RiPauseLine, RiPlayFill } from '@remixicon/react';
 import { toast } from 'sonner';
-import { useMutation, useQuery } from 'urql';
+import { useMutation } from 'urql';
 
-import { useEnvironment } from '@/components/Environments/environment-context';
 import { graphql } from '@/gql';
 
 type CurrentRunHandlingOption = {
@@ -26,25 +22,6 @@ const currentRunHandlingOptions = [
   { name: 'Cancel immediately', id: CURRENT_RUN_HANDLING_STRATEGY_CANCEL },
 ] as const;
 const defaultCurrentRunHandlingOption = currentRunHandlingOptions[0];
-
-const FunctionVersionNumberDocument = graphql(`
-  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {
-    workspace(id: $environmentID) {
-      workflow: workflowBySlug(slug: $slug) {
-        id
-        isPaused
-        name
-        archivedAt
-        current {
-          version
-        }
-        previous {
-          version
-        }
-      }
-    }
-  }
-`);
 
 const PauseFunctionDocument = graphql(`
   mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {
@@ -70,7 +47,7 @@ type PauseFunctionModalProps = {
   onClose: () => void;
 };
 
-function PauseFunctionModal({
+export function PauseFunctionModal({
   functionID,
   functionName,
   isPaused,
@@ -184,66 +161,5 @@ function PauseFunctionModal({
         </div>
       )}
     </AlertModal>
-  );
-}
-
-type PauseFunctionProps = {
-  functionSlug: string;
-  disabled: boolean;
-};
-
-export default function PauseFunctionButton({ functionSlug, disabled }: PauseFunctionProps) {
-  const [isPauseFunctionModalVisible, setIsPauseFunctionModalVisible] = useState<boolean>(false);
-  const environment = useEnvironment();
-
-  const [{ data: version, fetching: isFetchingVersions }] = useQuery({
-    query: FunctionVersionNumberDocument,
-    variables: {
-      environmentID: environment.id,
-      slug: functionSlug,
-    },
-  });
-
-  const fn = version?.workspace.workflow;
-
-  if (!fn) {
-    return null;
-  }
-
-  const { isPaused } = fn;
-
-  return (
-    <>
-      <Tooltip delayDuration={0}>
-        <TooltipTrigger asChild>
-          <span tabIndex={0}>
-            <Button
-              icon={
-                isPaused ? (
-                  <RiPlayFill className=" text-green-600" />
-                ) : (
-                  <RiPauseLine className=" text-amber-500" />
-                )
-              }
-              btnAction={() => setIsPauseFunctionModalVisible(true)}
-              disabled={disabled || isFetchingVersions}
-              label={isPaused ? 'Resume' : 'Pause'}
-            />
-          </span>
-        </TooltipTrigger>
-        <TooltipContent className="align-center rounded-md px-2 text-xs">
-          {isPaused
-            ? 'Begin running this function after a temporary pause'
-            : 'Temporarily stop a function from being run'}
-        </TooltipContent>
-      </Tooltip>
-      <PauseFunctionModal
-        functionID={fn.id}
-        functionName={fn.name}
-        isPaused={isPaused}
-        isOpen={isPauseFunctionModalVisible}
-        onClose={() => setIsPauseFunctionModalVisible(false)}
-      />
-    </>
   );
 }
