+++ ui/apps/dashboard/src/gql/gql.ts
@@ -46,15 +46,8 @@ const documents = {
     "\n  query GetSDKRequestMetrics(\n    $environmentID: ID!\n    $fnSlug: String!\n    $startTime: Time!\n    $endTime: Time!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $fnSlug) {\n        queued: metrics(opts: { name: \"sdk_req_scheduled_total\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n        started: metrics(opts: { name: \"sdk_req_started_total\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n\n        ended: metrics(opts: { name: \"sdk_req_ended_total\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n      }\n    }\n  }\n": types.GetSdkRequestMetricsDocument,
     "\n  query GetStepBacklogMetrics(\n    $environmentID: ID!\n    $fnSlug: String!\n    $startTime: Time!\n    $endTime: Time!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $fnSlug) {\n        scheduled: metrics(opts: { name: \"steps_scheduled\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n        sleeping: metrics(opts: { name: \"steps_sleeping\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n      }\n    }\n  }\n": types.GetStepBacklogMetricsDocument,
     "\n  query GetStepsRunningMetrics(\n    $environmentID: ID!\n    $fnSlug: String!\n    $startTime: Time!\n    $endTime: Time!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $fnSlug) {\n        running: metrics(opts: { name: \"steps_running\", from: $startTime, to: $endTime }) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n\n        concurrencyLimit: metrics(\n          opts: { name: \"concurrency_limit_reached_total\", from: $startTime, to: $endTime }\n        ) {\n          from\n          to\n          granularity\n          data {\n            bucket\n            value\n          }\n        }\n      }\n    }\n  }\n": types.GetStepsRunningMetricsDocument,
-    "\n  query GetFunctionArchival($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isArchived\n        name\n      }\n    }\n  }\n": types.GetFunctionArchivalDocument,
-    "\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n": types.CreateCancellationDocument,
-    "\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n": types.GetCancellationRunCountDocument,
-    "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n": types.GetFunctionVersionNumberDocument,
-    "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n": types.PauseFunctionDocument,
-    "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n": types.UnpauseFunctionDocument,
     "\n  mutation DeleteCancellation($envID: UUID!, $cancellationID: ULID!) {\n    deleteCancellation(envID: $envID, cancellationID: $cancellationID)\n  }\n": types.DeleteCancellationDocument,
     "\n  query GetFnCancellations($after: String, $envSlug: String!, $fnSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      fn: workflowBySlug(slug: $fnSlug) {\n        cancellations(after: $after) {\n          edges {\n            cursor\n            node {\n              createdAt\n              envID: environmentID\n              id\n              name\n              queuedAtMax\n              queuedAtMin\n            }\n          }\n          pageInfo {\n            hasNextPage\n          }\n        }\n      }\n    }\n  }\n": types.GetFnCancellationsDocument,
-    "\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n": types.InvokeFunctionDocument,
     "\n  mutation RerunFunctionRun($environmentID: ID!, $functionID: ID!, $functionRunID: ULID!) {\n    retryWorkflowRun(\n      input: { workspaceID: $environmentID, workflowID: $functionID }\n      workflowRunID: $functionRunID\n    ) {\n      id\n    }\n  }\n": types.RerunFunctionRunDocument,
     "\n  query GetHistoryItemOutput($envID: ID!, $functionID: ID!, $historyItemID: ULID!, $runID: ULID!) {\n    environment: workspace(id: $envID) {\n      function: workflow(id: $functionID) {\n        run(id: $runID) {\n          historyItemOutput(id: $historyItemID)\n        }\n      }\n    }\n  }\n": types.GetHistoryItemOutputDocument,
     "\n  query GetFunctionRunDetails($environmentID: ID!, $functionSlug: String!, $functionRunID: ULID!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        name\n        run(id: $functionRunID) {\n          batchID\n          canRerun\n          events {\n            id\n            name\n            payload: event\n            receivedAt\n          }\n          history {\n            attempt\n            cancel {\n              eventID\n              expression\n              userID\n            }\n            createdAt\n            functionVersion\n            groupID\n            id\n            sleep {\n              until\n            }\n            stepName\n            type\n            url\n            waitForEvent {\n              eventName\n              expression\n              timeout\n            }\n            waitResult {\n              eventID\n              timeout\n            }\n            invokeFunction {\n              eventID\n              functionID\n              correlationID\n              timeout\n            }\n            invokeFunctionResult {\n              eventID\n              timeout\n              runID\n            }\n          }\n          id\n          status\n          startedAt\n          endedAt\n          output\n          version: workflowVersion {\n            deploy {\n              id\n              createdAt\n            }\n            triggers {\n              eventName\n              schedule\n            }\n            url\n            validFrom\n            version\n          }\n        }\n        slug\n      }\n    }\n  }\n": types.GetFunctionRunDetailsDocument,
@@ -63,6 +56,7 @@ const documents = {
     "\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n": types.GetReplayRunCountsDocument,
     "\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n": types.CreateFunctionReplayDocument,
     "\n  query GetFunctionRunsCount(\n    $environmentID: ID!\n    $functionSlug: String!\n    $functionRunStatuses: [FunctionRunStatus!]\n    $timeRangeStart: Time!\n    $timeRangeEnd: Time!\n    $timeField: FunctionRunTimeField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        isPaused\n        runs: runsV2(\n          filter: {\n            status: $functionRunStatuses\n            lowerTime: $timeRangeStart\n            upperTime: $timeRangeEnd\n            timeField: $timeField\n          }\n        ) {\n          totalCount\n        }\n      }\n    }\n  }\n": types.GetFunctionRunsCountDocument,
+    "\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n": types.InvokeFunctionDocument,
     "\n  query GetReplays($environmentID: ID!, $functionSlug: String!) {\n    environment: workspace(id: $environmentID) {\n      id\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replays {\n          id\n          name\n          createdAt\n          endedAt\n          functionRunsScheduledCount\n        }\n      }\n    }\n  }\n": types.GetReplaysDocument,
     "\n  query GetFunctionPauseState($environmentID: ID!, $functionSlug: String!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        isPaused\n      }\n    }\n  }\n": types.GetFunctionPauseStateDocument,
     "\n  query GetRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            function {\n              name\n              slug\n            }\n            id\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n": types.GetRunsDocument,
@@ -98,6 +92,11 @@ const documents = {
     "\n  mutation UnarchiveEnvironment($id: ID!) {\n    unarchiveEnvironment(id: $id) {\n      id\n    }\n  }\n": types.UnarchiveEnvironmentDocument,
     "\n  mutation DisableEnvironmentAutoArchiveDocument($id: ID!) {\n    disableEnvironmentAutoArchive(id: $id) {\n      id\n    }\n  }\n": types.DisableEnvironmentAutoArchiveDocumentDocument,
     "\n  mutation EnableEnvironmentAutoArchive($id: ID!) {\n    enableEnvironmentAutoArchive(id: $id) {\n      id\n    }\n  }\n": types.EnableEnvironmentAutoArchiveDocument,
+    "\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n": types.CreateCancellationDocument,
+    "\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n": types.GetCancellationRunCountDocument,
+    "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n": types.GetFunctionVersionNumberDocument,
+    "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n": types.PauseFunctionDocument,
+    "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n": types.UnpauseFunctionDocument,
     "\n  query GetGlobalSearch($opts: SearchInput!) {\n    account {\n      search(opts: $opts) {\n        results {\n          env {\n            name\n            id\n            type\n            slug\n          }\n          kind\n          value {\n            ... on ArchivedEvent {\n              id\n              name\n            }\n            ... on FunctionRun {\n              id\n              functionID: workflowID\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetGlobalSearchDocument,
     "\n  query GetFunctionSlug($environmentID: ID!, $functionID: ID!) {\n    environment: workspace(id: $environmentID) {\n      function: workflow(id: $functionID) {\n        slug\n        name\n      }\n    }\n  }\n": types.GetFunctionSlugDocument,
     "\n  fragment TraceDetails on RunTraceSpan {\n    name\n    status\n    attempts\n    queuedAt\n    startedAt\n    endedAt\n    isRoot\n    outputID\n    spanID\n    stepOp\n    stepInfo {\n      __typename\n      ... on InvokeStepInfo {\n        triggeringEventID\n        functionID\n        timeout\n        returnEventID\n        runID\n        timedOut\n      }\n      ... on SleepStepInfo {\n        sleepUntil\n      }\n      ... on WaitForEventStepInfo {\n        eventName\n        expression\n        timeout\n        foundEventID\n        timedOut\n      }\n    }\n  }\n": types.TraceDetailsFragmentDoc,
@@ -267,39 +266,11 @@ export function graphql(source: "\n  query GetStepsRunningMetrics(\n    $environ
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query GetFunctionArchival($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isArchived\n        name\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFunctionArchival($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isArchived\n        name\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
 export function graphql(source: "\n  mutation DeleteCancellation($envID: UUID!, $cancellationID: ULID!) {\n    deleteCancellation(envID: $envID, cancellationID: $cancellationID)\n  }\n"): (typeof documents)["\n  mutation DeleteCancellation($envID: UUID!, $cancellationID: ULID!) {\n    deleteCancellation(envID: $envID, cancellationID: $cancellationID)\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query GetFnCancellations($after: String, $envSlug: String!, $fnSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      fn: workflowBySlug(slug: $fnSlug) {\n        cancellations(after: $after) {\n          edges {\n            cursor\n            node {\n              createdAt\n              envID: environmentID\n              id\n              name\n              queuedAtMax\n              queuedAtMin\n            }\n          }\n          pageInfo {\n            hasNextPage\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFnCancellations($after: String, $envSlug: String!, $fnSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      fn: workflowBySlug(slug: $fnSlug) {\n        cancellations(after: $after) {\n          edges {\n            cursor\n            node {\n              createdAt\n              envID: environmentID\n              id\n              name\n              queuedAtMax\n              queuedAtMin\n            }\n          }\n          pageInfo {\n            hasNextPage\n          }\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n"): (typeof documents)["\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -332,6 +303,10 @@ export function graphql(source: "\n  mutation CreateFunctionReplay(\n    $enviro
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query GetFunctionRunsCount(\n    $environmentID: ID!\n    $functionSlug: String!\n    $functionRunStatuses: [FunctionRunStatus!]\n    $timeRangeStart: Time!\n    $timeRangeEnd: Time!\n    $timeField: FunctionRunTimeField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        isPaused\n        runs: runsV2(\n          filter: {\n            status: $functionRunStatuses\n            lowerTime: $timeRangeStart\n            upperTime: $timeRangeEnd\n            timeField: $timeField\n          }\n        ) {\n          totalCount\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFunctionRunsCount(\n    $environmentID: ID!\n    $functionSlug: String!\n    $functionRunStatuses: [FunctionRunStatus!]\n    $timeRangeStart: Time!\n    $timeRangeEnd: Time!\n    $timeField: FunctionRunTimeField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        isPaused\n        runs: runsV2(\n          filter: {\n            status: $functionRunStatuses\n            lowerTime: $timeRangeStart\n            upperTime: $timeRangeEnd\n            timeField: $timeField\n          }\n        ) {\n          totalCount\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n"): (typeof documents)["\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug)\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -472,6 +447,26 @@ export function graphql(source: "\n  mutation DisableEnvironmentAutoArchiveDocum
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  mutation EnableEnvironmentAutoArchive($id: ID!) {\n    enableEnvironmentAutoArchive(id: $id) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation EnableEnvironmentAutoArchive($id: ID!) {\n    enableEnvironmentAutoArchive(id: $id) {\n      id\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation CreateCancellation($input: CreateCancellationInput!) {\n    createCancellation(input: $input) {\n      id\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetCancellationRunCount(\n    $envID: ID!\n    $functionSlug: String!\n    $queuedAtMin: Time\n    $queuedAtMax: Time!\n  ) {\n    environment: workspace(id: $envID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        cancellationRunCount(input: { queuedAtMin: $queuedAtMin, queuedAtMax: $queuedAtMax })\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFunctionVersionNumber($slug: String!, $environmentID: ID!) {\n    workspace(id: $environmentID) {\n      workflow: workflowBySlug(slug: $slug) {\n        id\n        isPaused\n        name\n        archivedAt\n        current {\n          version\n        }\n        previous {\n          version\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation PauseFunction($fnID: ID!, $cancelRunning: Boolean) {\n    pauseFunction(fnID: $fnID, cancelRunning: $cancelRunning) {\n      id\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation UnpauseFunction($fnID: ID!) {\n    unpauseFunction(fnID: $fnID) {\n      id\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
